backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
aiosqlite
httpx
pytest
pytest-asyncio